#171 refactoring spotRegister
